requirement.txt
asgiref==3.2.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.4
django-import-export==2.4.0
django-summernote==0.8.11.6
et-xmlfile==1.0.1
html5lib==1.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.0.0
PyPDF2==1.26.0
python-http-client==3.2.7
pytz==2019.3
PyYAML==5.3.1
reportlab==3.5.44
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
webencodings==0.5.1
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
